# api/api.yml
openapi: 3.0.2
info:
  title: "ISO8583 API"
  version: 0.0.1
  description: "Package github.com/moov-io/iso8583 implements a file reader and writer written in Go decorated with a HTTP API for creating, parsing, and validating financial transaction card originated interchange messaging. User can seed iso8583's specification file as json file, iso8583 message with several formsts (mesage binary, json, xml)\n
| Input      | Output     |\n
|------------|------------|\n
| JSON       | JSON       |\n
| XML        | XML        |\n
| MESSAGE    | MESSAGE    |\n
"
servers:
  - url: https://local.moov.io:8208/
    description: Local Testing
  - url: https://api.moov.io/
    description: Production
...
paths:
...
  /print:
    post:
      tags: ['iso8583 message']
      summary: Print iso8583 message with specific format
      description: Print iso8583 message with requested format.
      operationId: print
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                format:
                  type: string
                  description: print iso8583 type
                  default: json
                  example: iso8583
                  enum:
                    - json
                    - xml
                    - iso8583
                input:
                  type: string
                  description: iso8583 message file
                  format: binary
                  example:
                    mti: '0400'
                    bitmap: '1111001000111010010000000000000100001000010000011000001000000010'
                    elements:
                      '1': '0000000000000000000000000100000000000000000000000000000000000000'
                      '2': '1111111110000000000'
                      '3': '180000'
                      '4': '000000030000'
                      '7': 0109080646
                      '11': '100331'
                      '12': '001120'
                      '13': 0909
                      '15': 0809
                      '18': '9601'
                      '32': ''
                      '37': '600020000000'
                      '42': '00003430003948 '
                      '48': 08110012000004096565733200000003000001
                      '49': '360'
                      '55': 003317000394809080646510000000
                      '63': '132'
                      '90': 020000331609080645190000000020000000000000
                spec:
                  type: string
                  description: iso8583 message specification
                  format: binary
                  example:
                    elements:
                     '1':
                       Describe: b 64
                       Description: Second Bitmap
                     '2':
                       Describe: n..19
                       Description: Primary account number (PAN)
                     '3':
                       Describe: n 6
                       Description: Processing code
                     '4':
                       Describe: n 12
                       Description: 'Amount, transaction'
                     '5':
                       Describe: n 12
                       Description: 'Amount, settlement'
                     '6':
                       Describe: n 12
                       Description: 'Amount, cardholder billing'
                     '7':
                       Describe: n 10; MMDDhhmmss
                       Description: Transmission date & time
                     '8':
                       Describe: n 8
                       Description: 'Amount, cardholder billing fee'
                     '9':
                       Describe: n 8
                       Description: 'Conversion rate, settlement'
                     '10':
                       Describe: n 8
                       Description: 'Conversion rate, cardholder billing'
                     '11':
                       Describe: n 6
                       Description: System trace audit number (STAN)
                     '12':
                       Describe: n 6; hhmmss
                       Description: Local transaction time (hhmmss)
                     '13':
                       Describe: n 4; MMDD
                       Description: Local transaction date (MMDD)
                     '14':
                       Describe: n 4; YYMM
                       Description: Expiration date
                     '15':
                       Describe: n 4; MMDD
                       Description: Settlement date
                     '16':
                       Describe: n 4; MMDD
                       Description: Currency conversion date
                     '17':
                       Describe: n 4; MMDD
                       Description: Capture date
                     '18':
                       Describe: n 4
                       Description: 'Merchant type, or merchant category code'
                     '19':
                       Describe: n 3
                       Description: Acquiring institution (country code)
                     '20':
                       Describe: n 3
                       Description: PAN extended (country code)
                     '21':
                       Describe: n 3
                       Description: Forwarding institution (country code)
                     '22':
                       Describe: n 3
                       Description: Point of Sale (POS) entry mode
                     '23':
                       Describe: n 3
                       Description: Application PAN sequence number
                     '24':
                       Describe: n 3
                       Description: 'Function code (ISO 8583:1993), or network international identifier (NII)'
                     '25':
                       Describe: n 2
                       Description: Point of Sale (POS) condition code
                     '26':
                       Describe: n 2
                       Description: Point of Sale (POS) capture code
                     '27':
                       Describe: n 1
                       Description: Authorizing identification response length
                     '28':
                       Describe: x+n 8
                       Description: 'Amount, transaction fee'
                     '29':
                       Describe: x+n 8
                       Description: 'Amount, settlement fee'
                     '30':
                       Describe: x+n 8
                       Description: 'Amount, transaction processing fee'
                     '31':
                       Describe: x+n 8
                       Description: 'Amount, settlement processing fee'
                     '32':
                       Describe: n..11
                       Description: Acquiring institution identification code
                     '33':
                       Describe: n..11
                       Description: Forwarding institution identification code
                     '34':
                       Describe: ns..28
                       Description: 'Primary account number, extended'
                     '35':
                       Describe: z..37
                       Description: Track 2 data
                     '36':
                       Describe: n...104
                       Description: Track 3 data
                     '37':
                       Describe: an 12
                       Description: Retrieval reference number
                     '38':
                       Describe: an 6
                       Description: Authorization identification response
                     '39':
                       Describe: an 2
                       Description: Response code
                     '40':
                       Describe: an 3
                       Description: Service restriction code
                     '41':
                       Describe: ans 8
                       Description: Card acceptor terminal identification
                     '42':
                       Describe: ans 15
                       Description: Card acceptor identification code
                     '43':
                       Describe: ans 40
                       Description: 'Card acceptor name/location (1–23 street address, –36 city, –38 state, 39–40 country)'
                     '44':
                       Describe: an..25
                       Description: Additional response data
                     '45':
                       Describe: an..76
                       Description: Track 1 data
                     '46':
                       Describe: an...999
                       Description: Additional data (ISO)
                     '47':
                       Describe: an...999
                       Description: Additional data (national)
                     '48':
                       Describe: an...999
                       Description: Additional data (private)
                     '49':
                       Describe: n 3
                       Description: 'Currency code, transaction'
                     '50':
                       Describe: n 3
                       Description: 'Currency code, settlement'
                     '51':
                       Describe: n 3
                       Description: 'Currency code, cardholder billing'
                     '52':
                       Describe: b 64
                       Description: Personal identification number data
                     '53':
                       Describe: n 16
                       Description: Security related control information
                     '54':
                       Describe: an...120
                       Description: Additional amounts
                     '55':
                       Describe: ans...999
                       Description: ICC data – EMV having multiple tags
                     '56':
                       Describe: ans...999
                       Description: Reserved (ISO)
                     '57':
                       Describe: ans...999
                       Description: Reserved (national)
                     '58':
                       Describe: ans...999
                       Description: Reserved (national)
                     '59':
                       Describe: ans...999
                       Description: Reserved (national)
                     '60':
                       Describe: ans...999
                       Description: Reserved (national)
                     '61':
                       Describe: ans...999
                       Description: Reserved (private) (e.g. CVV2/service code   transactions)
                     '62':
                       Describe: ans...999
                       Description: 'Reserved (private) (e.g. transactions: invoice number, key exchange transactions: TPK key, etc.)'
                     '63':
                       Describe: ans...999
                       Description: Reserved (private)
                     '64':
                       Describe: b 64
                       Description: Message authentication code (MAC)
                     '65':
                       Describe: b 1
                       Description: Extended bitmap indicator
                     '66':
                       Describe: n 1
                       Description: Settlement code
                     '67':
                       Describe: n 2
                       Description: Extended payment code
                     '68':
                       Describe: n 3
                       Description: Receiving institution country code
                     '69':
                       Describe: n 3
                       Description: Settlement institution country code
                     '70':
                       Describe: n 3
                       Description: Network management information code
                     '71':
                       Describe: n 4
                       Description: Message number
                     '72':
                       Describe: n 4
                       Description: Last message's number
                     '73':
                       Describe: n 6; YYMMDD
                       Description: Action date (YYMMDD)
                     '74':
                       Describe: n 10
                       Description: Number of credits
                     '75':
                       Describe: n 10
                       Description: 'Credits, reversal number'
                     '76':
                       Describe: n 10
                       Description: Number of debits
                     '77':
                       Describe: n 10
                       Description: 'Debits, reversal number'
                     '78':
                       Describe: n 10
                       Description: Transfer number
                     '79':
                       Describe: n 10
                       Description: 'Transfer, reversal number'
                     '80':
                       Describe: n 10
                       Description: Number of inquiries
                     '81':
                       Describe: n 10
                       Description: Number of authorizations
                     '82':
                       Describe: n 12
                       Description: 'Credits, processing fee amount'
                     '83':
                       Describe: n 12
                       Description: 'Credits, transaction fee amount'
                     '84':
                       Describe: n 12
                       Description: 'Debits, processing fee amount'
                     '85':
                       Describe: n 12
                       Description: 'Debits, transaction fee amount'
                     '86':
                       Describe: n 16
                       Description: Total amount of credits
                     '87':
                       Describe: n 16
                       Description: 'Credits, reversal amount'
                     '88':
                       Describe: n 16
                       Description: Total amount of debits
                     '89':
                       Describe: n 16
                       Description: 'Debits, reversal amount'
                     '90':
                       Describe: n 42
                       Description: Original data elements
                     '91':
                       Describe: an 1
                       Description: File update code
                     '92':
                       Describe: an 2
                       Description: File security code
                     '93':
                       Describe: an 5
                       Description: Response indicator
                     '94':
                       Describe: an 7
                       Description: Service indicator
                     '95':
                       Describe: an 42
                       Description: Replacement amounts
                     '96':
                       Describe: b 64
                       Description: Message security code
                     '97':
                       Describe: x+n 16
                       Description: Net settlement amount
                     '98':
                       Describe: ans 25
                       Description: Payee
                     '99':
                       Describe: n..11
                       Description: Settlement institution identification code
                     '100':
                       Describe: n..11
                       Description: Receiving institution identification code
                     '101':
                       Describe: ans..17
                       Description: File name
                     '102':
                       Describe: ans..28
                       Description: Account identification 1
                     '103':
                       Describe: ans..28
                       Description: Account identification 2
                     '104':
                       Describe: ans...100
                       Description: Transaction description
                     '105':
                       Describe: ans...999
                       Description: Reserved for ISO use
                     '106':
                       Describe: ans...999
                       Description: Reserved for ISO use
                     '107':
                       Describe: ans...999
                       Description: Reserved for ISO use
                     '108':
                       Describe: ans...999
                       Description: Reserved for ISO use
                     '109':
                       Describe: ans...999
                       Description: Reserved for ISO use
                     '110':
                       Describe: ans...999
                       Description: Reserved for ISO use
                     '111':
                       Describe: ans...999
                       Description: Reserved for ISO use
                     '112':
                       Describe: ans...999
                       Description: Reserved for national use
                     '113':
                       Describe: ans...999
                       Description: Reserved for national use
                     '114':
                       Describe: ans...999
                       Description: Reserved for national use
                     '115':
                       Describe: ans...999
                       Description: Reserved for national use
                     '116':
                       Describe: ans...999
                       Description: Reserved for national use
                     '117':
                       Describe: ans...999
                       Description: Reserved for national use
                     '118':
                       Describe: ans...999
                       Description: Reserved for national use
                     '119':
                       Describe: ans...999
                       Description: Reserved for national use
                     '120':
                       Describe: ans...999
                       Description: Reserved for private use
                     '121':
                       Describe: ans...999
                       Description: Reserved for private use
                     '122':
                       Describe: ans...999
                       Description: Reserved for private use
                     '123':
                       Describe: ans...999
                       Description: Reserved for private use
                     '124':
                       Describe: ans...999
                       Description: Reserved for private use
                     '125':
                       Describe: ans...999
                       Description: Reserved for private use
                     '126':
                       Describe: ans...999
                       Description: Reserved for private use
                     '127':
                       Describe: ans...999
                       Description: Reserved for private use
                     '128':
                       Describe: b 64
                       Description: Message authentication code
                    encoding:
                     mti_enc: CHAR
                     bmp_enc: HEX
                     len_enc: CHAR
                     num_enc: CHAR
                     chr_enc: ASCII
                     bin_enc: HEX
                     trk_enc: EBCDIC
            encoding:
              file:
                contentType: text/plain
      responses:
        '200':
          description: successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                description: iso8583 message
                format: binary
                example: '0400F23A400108418202000000400000000019111111111000000000018000000000003000001090806461003310011200909080996010060002000000000003430003948 03808110012000004096565733200000003000001360030003317000394809080646510000000003132020000331609080645190000000020000000000000'
            application/json:
              schema:
                $ref: '#/components/schemas/ISOMessage'
            application/xml:
              schema:
                $ref: '#/components/schemas/ISOMessage'
        '400':
          description: bad request
          content:
            text/plain:
              schema:
                type: string
        '501':
          description: failed operation
          content:
            text/plain:
              schema:
                type: string
...
components:
...
  securitySchemes:
    GatewayAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT that comes from the gateway that validates against the gateways public RSA key
  schemas:
...
    ISOMessage:
      properties:
        mit:
          $ref: '#/components/schemas/Element'
        bitmap:
          $ref: '#/components/schemas/Element'
        data_elements:
          type: array
          items:
            $ref: '#/components/schemas/Element'
      required:
        - mit
        - bitmap
    Element:
      properties:
        type:
          type: string
        length:
          type: integer
          format: int32
        format:
          type: string
        encoding:
          type: string
        fixed:
          type: boolean
        length_encoding:
          type: string
        data_length:
          type: integer
          format: int32
        value:
          type: array
          items:
            type: string
            format: byte
...